requirements.txt
Django >= 2.1.4
djangorestframework >= 3.9.0
google-api-python-client >= 1.7.7
httplib2 >= 0.12.0
oauth2client >= 4.1.3
pytz >= 2018.7